Move pkg/cloud/aws/actuators/machine to pkg/actuators/machine
No need for cloud/aws in the machine actuator package path.
It's obvious from the context it's aws actuator.
