+++ pkg/actuators/machine/stubs.go
